requirements.txt
python-dotenv==1.0.1
supabase==2.4.6
resend==2.3.0
pydantic==2.8.2
python-dateutil==2.9.0.post0